requirements.txt
dbt-bigquery
pandas 
pandas-gbq 
google-auth 
google-cloud-bigquery
# apache-airflow-providers-google
